+++ test/test-runner.ts
@@ -282,6 +282,28 @@ test('exit-after-last-test', async (t) => {
     t.equal(testsuites.$.skipped, '0')
 })
 
+test('unhandled rejection', async (t) => {
+    const res = await runTest('./unhandled-rejection.js')
+    const testsuites = res.xml.testsuites
+
+    t.equal(res.exitCode, 1, 'test shall not pass')
+    t.equal(testsuites.$.tests, '2')
+    t.equal(testsuites.$.errors, '1')
+    t.equal(testsuites.$.failures, '0')
+    t.equal(testsuites.$.skipped, '0')
+})
+
+test('uncaught exception', async (t) => {
+    const res = await runTest('./uncaught-exception.js')
+    const testsuites = res.xml.testsuites
+
+    t.equal(res.exitCode, 1, 'test shall not pass')
+    t.equal(testsuites.$.tests, '2')
+    t.equal(testsuites.$.errors, '1')
+    t.equal(testsuites.$.failures, '0')
+    t.equal(testsuites.$.skipped, '0')
+})
+
 test('tryUntil', async (t) => {
     const res = await runTest('./tryUntil.js')
     const testsuites = res.xml.testsuites