+++ test/unhandled-rejection.ts
@@ -1,0 +1,10 @@
+import 'source-map-support/register'
+import { test } from '../index'
+
+test('unhandledRejection', (t) => {
+    setTimeout(
+        () => new Promise((_, reject) => reject(new Error('rejection'))),
+        100
+    )
+    t.pass('Test ended well')
+})